requirements.txt
httpx>=0.27.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
fastapi>=0.115.0
openpyxl>=3.1.0
